requirements.txt
apache-airflow==3.1.5
joblib==1.3.0
pandas==2.3.3
scikit_learn==1.8.0
